Add Deployments over the last year chart to Surge page
