+++ src/root/components/connected/emergencies-dash.js
@@ -6,6 +6,7 @@ import { DateTime } from 'luxon';
 import { Link } from 'react-router-dom';
 import { environment } from '#config';
 import Stats from '#components/emergencies/stats';
+import { TimeLineChart } from '#components/Charts';
 import EmergenciesMap from '#components/map/emergencies-map';
 import Progress from '#components/progress';
 import BlockLoading from '#components/block-loading';
@@ -46,43 +47,6 @@ class EmergenciesDash extends React.Component {
     );
   }
 
-  renderChart (data) {
-    const zone = 'utc';
-    const tickFormatter = (date) => DateTime.fromISO(date, {zone}).toFormat('MMM');
-    const contentFormatter = (payload) => {
-      if (!payload.payload[0]) { return null; }
-
-      const item = payload.payload[0].payload;
-      return (
-        <article className='chart-tooltip'>
-          <div className='chart-tooltip__contents'>
-            <dl>
-              <dd>
-                <Translate stringId='emergenciesDashDate' />
-              </dd>
-              <dt>{DateTime.fromISO(item.timespan, {zone}).toFormat('MMMM yyyy')}</dt>
-              <dd>
-                <Translate stringId='emergenciesDashTotal' />
-              </dd>
-              <dt>{item.count}</dt>
-            </dl>
-          </div>
-        </article>
-      );
-    };
-
-    return (
-      <ResponsiveContainer>
-        <LineChart data={data}>
-          <XAxis tickFormatter={tickFormatter} dataKey='timespan' axisLine={false} padding={{ left: 16, right: 16 }} />
-          <YAxis axisLine={false} tickLine={false} width={32} padding={{ bottom: 16 }} />
-          <Line type="monotone" dataKey="count" stroke="#f5333f" />
-          <Tooltip content={contentFormatter}/>
-        </LineChart>
-      </ResponsiveContainer>
-    );
-  }
-
   renderByMonth () {
     if (!this.props.aggregate.month) return null;
 
@@ -100,7 +64,7 @@ class EmergenciesDash extends React.Component {
             <h2 className='fold__title'><Translate stringId='emergenciesDashOverLastYear' /></h2>
           </figcaption>
           <div className='chart__container charts__container__rtl'>
-            {this.renderChart(data, 'month')}
+            <TimeLineChart data={data} />
           </div>
         </figure>
       </div>
